Moved store tests to store
