+++ tests/store/test_add_core.py
