+++ tests/store/test_add_schema.py
