+++ tests/store/test_store_models.py
@@ -18,16 +18,16 @@ def test_instantiate_bundle_no_name():
     # WHEN instantiating a bundle without a name
     bundle_obj = models.Bundle()
     # THEN assert it was instantiated as expected
-    assert bundle_obj.name == None
+    assert bundle_obj.name is None
     assert bundle_obj.versions == []
 
 
-def test_version_obj_full_path(project_dir, store, bundle_obj):
+def test_version_obj_full_path(project_dir, minimal_bundle_obj):
     """Test the to get the full path from a version object"""
     # GIVEN a bundle object
-    timestamp = bundle_obj.created_at
+    timestamp = minimal_bundle_obj.created_at
     # WHEN instantiating the version obj
-    version_obj = models.Version(created_at=timestamp, bundle=bundle_obj)
+    version_obj = models.Version(created_at=timestamp, bundle=minimal_bundle_obj)
     # THEN it should point to the correct folder
     root_path = version_obj.full_path
-    assert root_path == project_dir / bundle_obj.name / str(timestamp.date())
+    assert root_path == project_dir / minimal_bundle_obj.name / str(timestamp.date())
